Create README.md
This function LMPlot (Local Mode Plot) is a function that plots Polarization as a function of Temperature. It is an interactive function that asks the user to enter the initial and final value of the Temperature simulated. It will also ask the user to enter the name of the file in .dat format and it will return the plot of Polarization against Temperature. Note: This Function only works for Annealing Simulation Data.